requirements.txt
gTTS
flask
torch
transformers
pillow
streamlit

